Move serialization config to Hive-side
